Extend apply-message to deployment-state-helpers.
